Move BaseFilteredMetric to abstraction, use at CPU
