+++ robusta_krr/core/integrations/prometheus/metrics/memory_metric.py
@@ -1,60 +1,10 @@
-import datetime
-from typing import Any, Optional
-
 from robusta_krr.core.models.allocations import ResourceType
 
-from .base_metric import BaseMetricLoader, bind_metric
-
-PrometheusSeries = Any
+from .base_metric import bind_metric
+from .base_filtered_metric import BaseFilteredMetricLoader
 
 
 @bind_metric(ResourceType.Memory)
-class MemoryMetricLoader(BaseMetricLoader):
-    @staticmethod
-    def get_target_name(series: PrometheusSeries) -> Optional[str]:
-        for label in ["container", "pod", "node"]:
-            if label in series["metric"]:
-                return series["metric"][label]
-        return None
-
-    @staticmethod
-    def filter_prom_jobs_results(
-        series_list_result: list[PrometheusSeries],
-    ) -> list[PrometheusSeries]:
-        """
-        Because there might be multiple metrics with the same name, we need to filter them out.
-
-        :param series_list_result: list of PrometheusSeries
-        """
-
-        if len(series_list_result) == 1:
-            return series_list_result
-
-        target_names = {
-            MemoryMetricLoader.get_target_name(series)
-            for series in series_list_result
-            if MemoryMetricLoader.get_target_name(series)
-        }
-        return_list: list[PrometheusSeries] = []
-
-        # takes kubelet job if exists, return first job alphabetically if it doesn't
-        for target_name in target_names:
-            relevant_series = [
-                series for series in series_list_result if MemoryMetricLoader.get_target_name(series) == target_name
-            ]
-            relevant_kubelet_metric = [series for series in relevant_series if series["metric"].get("job") == "kubelet"]
-            if len(relevant_kubelet_metric) == 1:
-                return_list.append(relevant_kubelet_metric[0])
-                continue
-            sorted_relevant_series = sorted(relevant_series, key=lambda s: s["metric"].get("job"), reverse=False)
-            return_list.append(sorted_relevant_series[0])
-        return return_list
-
-    async def query_prometheus(
-        self, query: str, start_time: datetime.datetime, end_time: datetime.datetime, step: datetime.timedelta
-    ) -> list[PrometheusSeries]:
-        result = await super().query_prometheus(query, start_time, end_time, step)
-        return self.filter_prom_jobs_results(result)
-
+class MemoryMetricLoader(BaseFilteredMetricLoader):
     def get_query(self, namespace: str, pod: str, container: str) -> str:
         return f'sum(container_memory_working_set_bytes{{image!="", namespace="{namespace}", pod="{pod}", container="{container}"}}) by (container, pod, job)'
